+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/ext/CreditCard.kt
@@ -5,11 +5,8 @@
 package mozilla.components.browser.engine.gecko.ext
 
 import mozilla.components.concept.engine.prompt.CreditCard
+import mozilla.components.support.utils.creditCardIIN
 import org.mozilla.geckoview.Autocomplete
-
-// Placeholder for the card type. This will be replaced when we can identify the card type.
-// This is dependent on https://github.com/mozilla-mobile/android-components/issues/9813.
-private const val CARD_TYPE_PLACEHOLDER = ""
 
 /**
  * Converts a GeckoView [Autocomplete.CreditCard] to an Android Components [CreditCard].
@@ -20,7 +17,7 @@ fun Autocomplete.CreditCard.toCreditCard() = CreditCard(
     number = number,
     expiryMonth = expirationMonth,
     expiryYear = expirationYear,
-    cardType = CARD_TYPE_PLACEHOLDER
+    cardType = number.creditCardIIN()?.creditCardIssuerNetwork?.name ?: ""
 )
 
 /**